Log replication, some remarks about leader election, some rework may be needed

+++ src/main/scala/my/lab/raft/LogEntry.scala
@@ -3,4 +3,4 @@ package my.lab.raft
 
 trait Command
 
-case class LogEntry(val command: Command, val term: Int)
+case class LogEntry(val command: Command, val term: Long)
